# docs/swagger.yaml
basePath: /
definitions:
...
  models.Extension:
    properties:
      base_url:
        description: Implementation details
        type: string
      capabilities:
        items:
          $ref: '#/definitions/models.ExtensionCapability'
        type: array
      created_at:
        type: string
      description:
        type: string
      icon:
        description: URL or base64
        type: string
      id:
        type: string
      installed:
        type: boolean
      installed_at:
        type: string
      name:
        type: string
      permissions:
        description: Required scopes
        items:
          type: string
        type: array
      publisher:
        type: string
      settings:
        additionalProperties: true
        description: App-specific config
        type: object
      status:
        $ref: '#/definitions/models.ExtensionStatus'
      updated_at:
        type: string
      version:
        type: string
      webhook_url:
        type: string
      widget_url:
        description: iFrame URL
        type: string
    type: object
  models.ExtensionCapability:
    enum:
    - widget
    - webhook
    - api
    type: string
    x-enum-comments:
      CapabilityAPI: Data integration
      CapabilityWebhook: Real-time events
      CapabilityWidget: UI integration
    x-enum-descriptions:
    - UI integration
    - Real-time events
    - Data integration
    x-enum-varnames:
    - CapabilityWidget
    - CapabilityWebhook
    - CapabilityAPI
  models.ExtensionStatus:
    enum:
    - active
    - inactive
    - pending
    type: string
    x-enum-varnames:
    - ExtensionStatusActive
    - ExtensionStatusInactive
    - ExtensionStatusPending
  models.FieldType:
    enum:
    - text
    - number
    - date
    - boolean
    - lookup
    - email
    - phone
    - file
    - url
    - textarea
    - select
    - multiselect
    - currency
    type: string
    x-enum-varnames:
    - FieldTypeText
    - FieldTypeNumber
    - FieldTypeDate
    - FieldTypeBoolean
    - FieldTypeLookup
    - FieldTypeEmail
    - FieldTypePhone
    - FieldTypeFile
    - FieldTypeURL
    - FieldTypeTextArea
    - FieldTypeSelect
    - FieldTypeMultiSelect
    - FieldTypeCurrency
...
  models.LookupDef:
    properties:
      display_field:
        description: Target Field to display in UI
        type: string
      module:
        description: Target Module Name
        type: string
      value_field:
        description: Target Field to store (usually _id)
        type: string
    type: object
  models.Module:
    properties:
      created_at:
        type: string
      fields:
        items:
          $ref: '#/definitions/models.ModuleField'
        type: array
      id:
        type: string
      is_system:
        description: If true, cannot be deleted
        type: boolean
      label:
        type: string
      name:
        description: Unique Identifier (e.g., "leads", "deals")
        type: string
      updated_at:
        type: string
    type: object
  models.ModuleField:
    properties:
      is_system:
        type: boolean
      label:
        type: string
      lookup:
        $ref: '#/definitions/models.LookupDef'
      name:
        type: string
      options:
        description: For Select/MultiSelect
        items:
          $ref: '#/definitions/models.SelectOptions'
        type: array
      required:
        type: boolean
      type:
        $ref: '#/definitions/models.FieldType'
    type: object
...
  models.SelectOptions:
    properties:
      label:
        type: string
      value:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: This is a sample server using Fiber, Uber Fx, and GORM.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Microservice Demo API
  version: "1.0"
paths:
...
  /api/admin/handle-webhook:
    post:
      consumes:
      - application/json
      description: Receive and print webhook payload
      parameters:
      - description: Webhook Payload
        in: body
        name: payload
        required: true
        schema:
          type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Webhook handled successfully!
          schema:
            type: string
      summary: Handle Webhook
      tags:
      - admin
...
  /api/extensions:
    get:
      consumes:
      - application/json
      description: Get all available marketplace extensions
      parameters:
      - description: Only show installed extensions
        in: query
        name: installed
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Extension'
            type: array
      summary: List Extensions
      tags:
      - extensions
    post:
      consumes:
      - application/json
      description: Create a new marketplace extension (Admin only)
      parameters:
      - description: Extension Data
        in: body
        name: extension
        required: true
        schema:
          $ref: '#/definitions/models.Extension'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Extension'
      summary: Create Extension
      tags:
      - extensions
  /api/extensions/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific extension by ID
      parameters:
      - description: Extension ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Extension'
      summary: Get Extension
      tags:
      - extensions
  /api/extensions/{id}/install:
    post:
      consumes:
      - application/json
      description: Install an extension by ID
      parameters:
      - description: Extension ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Install Extension
      tags:
      - extensions
  /api/extensions/{id}/uninstall:
    post:
      consumes:
      - application/json
      description: Uninstall an extension by ID
      parameters:
      - description: Extension ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Uninstall Extension
      tags:
      - extensions
...
  /api/modules:
    get:
      description: Get a list of all defined modules
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Module'
            type: array
      summary: List all modules
      tags:
      - modules
    post:
      consumes:
      - application/json
      description: Create a dynamic module with fields
      parameters:
      - description: Module Definition
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Module'
      produces:
      - application/json
      responses:
        "201":
          description: Module created
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Create a new module definition
      tags:
      - modules
...
  /api/modules/{name}:
    delete:
      consumes:
      - application/json
      description: Delete a module and its definition
      parameters:
      - description: Module Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Delete a module
      tags:
      - modules
    get:
      description: Get specific module definition including fields
      parameters:
      - description: Module Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Module'
        "404":
          description: Module not found
          schema:
            type: string
      summary: Get a module by name
      tags:
      - modules
    put:
      consumes:
      - application/json
      description: Update schema for a module
      parameters:
      - description: Module Name
        in: path
        name: name
        required: true
        type: string
      - description: Module Definition
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Module'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Update a module definition
      tags:
      - modules
...
  /api/notifications:
    get:
      consumes:
      - application/json
      description: Get a list of notifications for the current user
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List notifications
      tags:
      - Notifications
  /api/notifications/{id}/read:
    put:
      consumes:
      - application/json
      description: Mark a specific notification as read
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Mark notification as read
      tags:
      - Notifications
  /api/notifications/mark-all-read:
    post:
      consumes:
      - application/json
      description: Mark all notifications for the current user as read
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Mark all as read
      tags:
      - Notifications
  /api/notifications/unread-count:
    get:
      consumes:
      - application/json
      description: Get the number of unread notifications
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get unread count
      tags:
      - Notifications
...
swagger: "2.0"